Move all types into a separate file
