+++ client/mass/types/mass.ts
@@ -2,7 +2,7 @@ import type { RxAppApi } from '../../types/rx'
 import type { Menu } from '#dom'
 import type { Elem } from '../../types/d3'
 import type { ClientCopyGenome } from 'types/global'
-import type { BoxPlotSettings } from '../../plots/boxplot/BoxPlot'
+import type { BoxPlotSettings } from '../../plots/boxplot/BoxPlotTypes'
 import type { CorrVolcanoSettings } from '#plots/corrVolcano/CorrelationVolcanoTypes.ts'
 import type { VolcanoSettings } from '../../plots/volcano/VolcanoTypes.ts'
 //import { TermWrapper } from '#types'
