+++ client/plots/boxplot/BoxPlot.ts
@@ -4,71 +4,13 @@ import { controlsInit, term0_term2_defaultQ, renderTerm1Label } from '../control
 import { RxComponentInner } from '../../types/rx.d'
 import { plotColor } from '#shared/common.js'
 import { Menu, getMaxLabelWidth } from '#dom'
-import type { Div, Elem, SvgG, SvgSvg, SvgText } from '../../types/d3'
+import type { Elem } from '../../types/d3'
 import type { BasePlotConfig, MassAppApi, MassState, PlotConfig } from '#mass/types/mass'
+import type { TdbBoxPlotOpts, BoxPlotSettings, BoxPlotDom } from './BoxPlotTypes'
 import { Model } from './model/Model'
 import { ViewModel } from './viewModel/ViewModel'
 import { View } from './view/View'
 import { BoxPlotInteractions } from './interactions/BoxPlotInteractions'
-
-/** Opts sent from mass */
-type TdbBoxPlotOpts = {
-	holder: Elem
-	controls?: Elem
-	header?: Elem
-	numericEditMenuVersion?: string[]
-}
-
-/** User controlled settings. Some settings are calculated based on
- * the number of boxplots */
-export type BoxPlotSettings = {
-	/** Width of the boxplots and scale, excluding labels */
-	boxplotWidth: number
-	/** Default is common plot color.  */
-	color: string
-	/** Toggle between a white and black background */
-	darkMode: boolean
-	/** Padding between the left hand label and boxplot */
-	labelPad: number
-	/** Toggle between a linear and log scale
-	 * When true, renders a log scale. Default is false */
-	isLogScale: boolean
-	/** Toggle between vertical and horizontal orientation.
-	 * The default is false */
-	isVertical: boolean
-	/** If true, order box plots from smallest to largest median value
-	 * Default is by alphanumeric order or by bin
-	 * May change this later to `orderBy` if more options arise */
-	orderByMedian: boolean
-	/** Height of individual boxplots */
-	rowHeight: number
-	/** Space between the boxplots */
-	rowSpace: number
-}
-
-/** Descriptions of the dom elements for the box plot */
-export type BoxPlotDom = {
-	/** Div for boxplots below the scale */
-	boxplots: SvgG
-	/** Controls div for the hamburger menu */
-	controls: Elem
-	/** Main div */
-	div: Div
-	/** Error messages */
-	error: Div
-	/** Sandbox header */
-	header?: Elem
-	/** Legend */
-	legend: Div
-	/** Displays the term1 name as the plot title */
-	plotTitle: SvgText
-	/** Main svg holder */
-	svg: SvgSvg
-	/** Y-axis shown above the boxplots */
-	axis: any
-	/** box plot tooltip (e.g. over the outliers) */
-	tip: Menu
-}
 
 class TdbBoxplot extends RxComponentInner {
 	readonly type = 'boxplot'
