+++ client/plots/boxplot/view/View.ts
@@ -3,8 +3,7 @@ import { drawBoxplot, Menu } from '#dom'
 import { scaleLinear, scaleLog } from 'd3-scale'
 import { axisstyle } from '#src/client'
 import { axisTop, axisLeft } from 'd3-axis'
-import type { BoxPlotDom, BoxPlotSettings } from '../BoxPlot'
-import type { ViewData, PlotDimensions } from '../viewModel/ViewModel'
+import type { BoxPlotDom, BoxPlotSettings, ViewData, PlotDimensions } from '../BoxPlotTypes'
 import type { MassAppApi } from '#mass/types/mass'
 import type { BoxPlotInteractions } from '../interactions/BoxPlotInteractions'
 import type { RenderedPlot } from './RenderedPlot'
